Make `dtype` of the DRR settable (#101)

+++ test/test_xray.py
@@ -5,9 +5,25 @@ from diffdrr import Detector
 
 def test_Detector():
     try:
-        detector = Detector(100, 100, 7.5e-2, 7.5e-2, n_subsample=None, device="cuda")
+        detector = Detector(
+            100,
+            100,
+            7.5e-2,
+            7.5e-2,
+            n_subsample=None,
+            dtype=torch.float32,
+            device="cuda",
+        )
     except ValueError:
-        detector = Detector(100, 100, 7.5e-2, 7.5e-2, n_subsample=None, device="cpu")
+        detector = Detector(
+            100,
+            100,
+            7.5e-2,
+            7.5e-2,
+            n_subsample=None,
+            dtype=torch.float32,
+            device="cpu",
+        )
     sdr = torch.tensor([200.0]).to(detector.device)
     rotations = torch.rand(4, 3).to(detector.device)
     translations = torch.rand(4, 3).to(detector.device)
